Started refactoring on mobile responsiveness
